Implement header signing for V4
It turns out we hardly need any of the V2 code in V4, so most of that has been moved into V2Signer.cs. All we need is the trimming part.
